+++ modules/validators/builtin/cpp/mustBeVectorBuiltin.cpp
@@ -23,26 +23,52 @@
 // License along with this program. If not, see <http://www.gnu.org/licenses/>.
 // LICENCE_BLOCK_END
 //=============================================================================
-#include "mustBeTextScalarBuiltin.hpp"
+#include "mustBeVectorBuiltin.hpp"
 #include "ValidatorsInternal.hpp"
 //=============================================================================
 using namespace Nelson;
 //=============================================================================
 ArrayOfVector
-Nelson::ValidatorsGateway::mustBeTextScalarBuiltin(int nLhs, const ArrayOfVector& argIn)
+Nelson::ValidatorsGateway::mustBeVectorBuiltin(int nLhs, const ArrayOfVector& argIn)
 {
     ArrayOfVector retval;
     nargoutcheck(nLhs, 0, 0);
-    nargincheck(argIn, 1, 2);
+    nargincheck(argIn, 1, 3);
+    bool allowsAllEmpties = false;
     int argPos = -1;
-    if (argIn.size() == 2) {
+
+    switch (argIn.size()) {
+    case 2: {
+        if (argIn[1].isNumeric() && argIn[1].isScalar()) {
+            ArrayOf param2 = argIn[1];
+            argPos = param2.getContentAsInteger32Scalar();
+            if (argPos < 1) {
+                Error(_W("The last argument must be a positive integer."));
+            }
+        } else if (argIn[1].isRowVectorCharacterArray()
+            || (argIn[1].isStringArray() && argIn[1].isScalar())) {
+            std::wstring flag = argIn[1].getContentAsWideString();
+            if (flag != L"allows-all-empties") {
+                Error(_W("The last argument must be a positive integer or 'allows-all-empties'."));
+            }
+            allowsAllEmpties = true;
+        } else {
+            Error(_W("The last argument must be a positive integer or 'allows-all-empties'."));
+        }
+    } break;
+    case 3: {
+        std::wstring flag = argIn[1].getContentAsWideString();
+        if (flag != L"allows-all-empties") {
+            Error(_W("Second argument must be a 'allows-all-empties'."));
+        }
         ArrayOf param2 = argIn[1];
         argPos = param2.getContentAsInteger32Scalar();
         if (argPos < 1) {
             Error(_W("The last argument must be a positive integer."));
         }
+    } break;
     }
-    mustBeTextScalar(argIn[0], argPos, true);
+    mustBeVector(argIn[0], allowsAllEmpties, argPos, true);
     return retval;
 }
 //=============================================================================