+++ modules/validators/src/include/Validators.hpp
@@ -51,6 +51,8 @@ mustBeTextScalar(const ArrayOfVector& args, int argPosition);
 NLSVALIDATORS_IMPEXP void
 mustBeFolder(const ArrayOfVector& args, int argPosition);
 //=============================================================================
-
+NLSVALIDATORS_IMPEXP void
+mustBeVector(const ArrayOfVector& args, bool allowsAllEmpties, int argPosition);
+//=============================================================================
 }
 //=============================================================================